pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domain-netutils-monitor"
version = "0.1.0"
description = "Comprehensive domain and network utilities monitoring agent for WHOIS, SSL, DNS, HTTP, security records, and RBL checks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Domain Monitor Team", email = "team@example.com"}
]
maintainers = [
    {name = "Domain Monitor Team", email = "team@example.com"}
]
keywords = ["domain", "monitoring", "dns", "ssl", "whois", "security", "rbl", "dmarc", "spf", "dkim"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Networking",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: Security",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "dnspython>=2.0",
    "python-whois>=0.8",
    "aiohttp>=3.8",
    "pyOpenSSL>=23.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
domain-monitor = "domain_monitor.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=domain_monitor --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"
